Move parse_note_json into the note app
The helper decodes NoteContent.json values and has nothing to do with
registered reports; it only lived in registered_report_note_metadata
because that code needed it first. Relocate it next to NoteContent and
point all importers at note.models, so the registered-report module can
be deleted wholesale later.
