+++ src/note/models.py
@@ -1,8 +1,9 @@
-from note.related_models.note_model import Note, NoteContent
+from note.related_models.note_model import Note, NoteContent, parse_note_json
 from note.related_models.note_template_model import NoteTemplate
 
 __all__ = [
     "Note",
     "NoteContent",
     "NoteTemplate",
+    "parse_note_json",
 ]
